% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritance.R
\docType{data}
\name{analysisResults}
\alias{analysisResults}
\title{All observed and permutation results formatted in a
\code{methylInheritanceResults} class
object.}
\format{A \code{list} TODO}
\usage{
data(analysisResults)
}
\value{
A \code{list} TODO
}
\description{
TODO
}
\examples{

## Loading dataset containing all results
data(analysisResults)

## Extract information for the intersection between conserved differentially
## methylated sites (type = sites) between the intersection of 2
## generations (inter = i2): F1 and F2 (position = 1)
formatForGraph(allResults = analysisResults, type = "sites", inter="i2", 1)

}
\seealso{
\itemize{
    \item \code{\link{formatForGraph}} {for TODO}
}
}
\keyword{datasets}

