% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceInternalMethods.R
\name{createOutputDir}
\alias{createOutputDir}
\title{TODO}
\usage{
createOutputDir(output_dir, nbrGenerations, doingSites = TRUE,
  doingTiles = FALSE)
}
\arguments{
\item{output_dir}{a string of \code{character}, TODO}

\item{nbrGenerations}{a positive \code{integer}, TODO}

\item{doingSites}{a \code{logical}, TODO}

\item{doingTiles}{a \code{logical}, TODO}
}
\value{
\code{0} when all directories are created without problem.
}
\description{
TODO
}
\examples{

## TODO

}
\author{
Astrid Deschenes
}
\keyword{internal}

