% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceMethods.R
\name{doAnalysis}
\alias{doAnalysis}
\title{TODO}
\usage{
doAnalysis(realAnalysis_output_dir, permutations_output_dir,
  doingSites = TRUE)
}
\arguments{
\item{realAnalysis_output_dir}{a string, the name of the directory that
will contain the results of the permutation. If the directory does not
exist, it will be created.}

\item{permutations_output_dir}{a string, the name of the directory that
will contain the results of the permutation. If the directory does not
exist, it will be created.}
}
\value{
TODO
}
\description{
TODO
}
\examples{

##TODO

}
\author{
Astrid Deschenes, Pascal Belleau
}

