% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceMethods.R
\name{extractData}
\alias{extractData}
\title{TODO}
\usage{
extractData(realAnalysis_output_dir, permutations_output_dir,
  doingSites = TRUE, doingTiles = FALSE)
}
\arguments{
\item{realAnalysis_output_dir}{a string, the name of the directory that
will contain the results of the permutation. If the directory does not
exist, it will be created.}

\item{permutations_output_dir}{a string, the name of the directory that
will contain the results of the permutation. If the directory does not
exist, it will be created.}

\item{doingSites}{a \code{logical}, TODO}

\item{doingTiles}{a \code{logical}, TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

##TODO

}
\author{
Astrid Deschenes, Pascal Belleau
}

