% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceInternalMethods.R
\name{extractDataFromFile}
\alias{extractDataFromFile}
\title{TODO}
\usage{
extractDataFromFile(directory, elementType = c("SITES", "TILES"))
}
\arguments{
\item{directory}{a string of \code{character}, TODO}

\item{elementType}{a sting of \code{character}, TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## TODO

}
\author{
Astrid Deschenes
}
\keyword{internal}

