% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceMethods.R
\name{formatForGraph}
\alias{formatForGraph}
\title{TODO}
\usage{
formatForGraph(analysisandPermutationResults, type = c("sites", "tiles"),
  inter = c("i2", "iAll"), position)
}
\arguments{
\item{analysisandPermutationResults}{TODO}

\item{type}{One of the "sites" or "tiles" strings. Specifies the type
of differentially methylated elements should be returned. For
retrieving differentially methylated bases type="sites"; for
differentially methylated regions type="tiles". Default: "sites".}

\item{inter}{TODO}

\item{position}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## Loading dataset containing all results
data(analysisResults)

## Extract information for the intersection between conserved differentially
## methylated sites (type = sites) between the intersection of 2
## generations (inter = i2): F2 and F3 (position = 2)
formatForGraph(analysisandPermutationResults = analysisResults,
type = "sites", inter="i2", 2)

}
\author{
Astrid Deschenes, Pascal Belleau
}

