% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceInternalMethods.R
\name{formatInputMethylData}
\alias{formatInputMethylData}
\title{TODO}
\usage{
formatInputMethylData(methylKitData)
}
\arguments{
\item{methylKitData}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## TODO

}
\author{
Astrid Deschenes, Pascal Belleau
}
\keyword{internal}
