% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceInternalMethods.R
\name{getSampleNameFromFileName}
\alias{getSampleNameFromFileName}
\title{Extract sample name from file name}
\usage{
getSampleNameFromFileName(fileName)
}
\arguments{
\item{fileName}{TODO}
}
\value{
TODO
}
\description{
Extract sample name from file name
}
\examples{

##TODO

}
\author{
Astrid Deschenes, Pascal Belleau
}
\keyword{internal}

