% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceInternalMethods.R
\name{interGeneration}
\alias{interGeneration}
\title{TODO}
\usage{
interGeneration(resultAllGenGR)
}
\arguments{
\item{resultAllGenGR}{\code{GRanges} from \code{getGRangesFromMethylDiff}, TODO}
}
\value{
\code{list} with 2 elements
        i2 list of intersection G1 and G2, G2 and G3, ...
        iAll list of intersection G1 and G2 and G3,
             G1 and G2 and G3 and G4, ...
        TODO
}
\description{
TODO
}
\examples{

## Load permutation results on sites
permutationResultsFile <- dir(system.file("extdata",
package = "methylInheritance"), pattern = "permutationResultsForSites.RDS",
full.names = TRUE)
permutationResults <- readRDS(permutationResultsFile)

## Transform result to GRanges
resultsGR <- methylInheritance:::getGRangesFromMethylDiff(methDiff =
permutationResults, pDiff = 10, qvalue = 0.01, type = "hyper")

## Extract inter generational conserved sites
conservedSitesGR <- methylInheritance:::interGeneration(resultsGR)

}
\author{
Pascal Belleau
}
\keyword{internal}

