% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceInternalMethods.R
\name{isInterGenerationResults}
\alias{isInterGenerationResults}
\title{Verify if a specific file containing intergenerational results
exists or not.}
\usage{
isInterGenerationResults(outputDir, permutationID, type = c("sites",
  "tiles"))
}
\arguments{
\item{outputDir}{a string of \code{character}, the name of the directory
that will contain
the results of the permutation. The name should end with a slash. The
directory should already exists.}

\item{permutationID}{an \code{integer}, the identifiant of the permutation.
When the \code{permutationID} = \code{0}, the results are considered as the
observed results and are saved in a file with the "_observed_results.RDS"
extension. When the \code{permutationID} != \code{0}, the results are
considered as permutation results and are saved in a file with the
"_permutation_{permutationID}.RDS" extension.}

\item{type}{One of the \code{"sites"} or \code{"tiles"} strings. Specifies
the type of differentially methylated elements should be saved.
Default: \code{"sites"}.}
}
\value{
\code{TRUE} when file present; otherwise \code{FALSE}.
}
\description{
Verify if a specific file containing intergenerational results
exists or not.
}
\examples{

## Get the name of the directory where the file is stored
filesDir <- system.file("extdata", "TEST", package="methylInheritance")

## Verify that DMS intergenerational results for the observed data exists
methylInheritance:::isInterGenerationResults(outputDir =
    paste0(filesDir, "/"), 0, "sites")

}
\author{
Astrid Deschenes, Pascal Belleau
}
\keyword{internal}
