% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceMethods.R
\name{loadAllPermutationRDS}
\alias{loadAllPermutationRDS}
\title{TODO}
\usage{
loadAllPermutationRDS(analysisResultsDIR, permutationResultsDIR,
  doingSites = TRUE, doingTiles = FALSE)
}
\arguments{
\item{analysisResultsDIR}{TODO}

\item{permutationResultsDIR}{TODO}

\item{doingSites}{TODO}

\item{doingTiles}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## TODO

}
\author{
Astrid Deschenes, Pascal Belleau
}

