% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritance.R
\docType{package}
\name{methylInheritance-package}
\alias{methylInheritance-package}
\alias{methylInheritance}
\title{methylInheritance: Permutation-Based Analysis associating Conserved
Differentially Methylated Elements from One Generation to the Next to
a Treatment Effect}
\description{
This package does a permutation analysis, based on Monte Carlo sampling,
for testing the hypothesis that the number of conserved differentially
methylated elements (sites or tiles), between
several generations, is associated to an effect inherited from a treatment
and that stochastic effect can be dismissed.
}
\seealso{
\itemize{
    \item \code{\link{runPermutation}} {for running a
    permutation analysis, and optionally an observation analysis, on a
    specified multi-generational dataset}
    \item \code{\link{runObservation}} {for running an
    observation analysis on a specified multi-generational dataset}
}
}
\author{
Astrid Deschênes,
Pascal Belleau and
Arnaud Droit

Maintainer:
Astrid Deschenes <adeschen@hotmail.com>
}
\keyword{package}
