% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.methylInheritanceAllResults.R
\name{print.methylInheritanceAllResults}
\alias{print.methylInheritanceAllResults}
\title{Print a methylInheritanceAllResults object}
\usage{
\method{print}{methylInheritanceAllResults}(x, ...)
}
\arguments{
\item{x}{the output object from \code{mergePermutationAndObservation}
function, \code{runPermutationUsingRDSFile} function (when
\code{runObservationAnalysis} = \code{TRUE} and
\code{runPermutationUsingMethylKitInfo} function (when
\code{runObservationAnalysis} = \code{TRUE} to be printed}

\item{\ldots}{arguments passed to or from other methods}
}
\description{
Print a methylInheritanceAllResults object
}

