% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceMethods.R
\name{runObservationUsingRDSFile}
\alias{runObservationUsingRDSFile}
\title{Run a differentially methylation analysis on each generation
present in a dataset}
\usage{
runObservationUsingRDSFile(methylKitRDSFile, type = c("both", "sites",
  "tiles"), outputDir = NULL, nbrPermutations = 0, nbrCores = 1,
  nbrCoresDiffMeth = 1, minReads = 10, minMethDiff = 10, qvalue = 0.01,
  maxPercReads = 99.9, destrand = FALSE, minCovBasesForTiles = 0,
  tileSize = 1000, stepSize = 1000, vSeed = -1)
}
\arguments{
\item{methylKitRDSFile}{a string, the name of the RDS file containing the
methylKit objet used for the permutation analysis. The RDS file must
hold a \code{list} of \code{methylRawList} entries, each
\code{methylRawList} contains all the \code{methylRaw} entries related to
one generation (first entry = first generation, second entry = second
generation, etc..). The number of generations must correspond to the number
of entries in the \code{methylKitInfo}.At least 2 generations
must be present to do a permutation analysis. More information can be found
in the methylKit package.}

\item{type}{One of the "sites","tiles" or "both" strings. Specifies the type
of differentially methylated elements should be returned. For
retrieving differentially methylated bases type="sites"; for
differentially methylated regions type="tiles". Default: "both".}

\item{outputDir}{a string, the name of the directory that will contain
the results of the permutation or \code{NULL}. If the directory does not
exist, it will be created. When \code{NULL}, the results of the permutation
are not saved. Default: \code{NULL}.}

\item{nbrPermutations, }{a positive \code{integer}, the total number of
permutations that is going to be done. Default: \code{1000}.}

\item{nbrCores}{a positive \code{integer}, the number of cores to use when
processing the analysis. Default: \code{1} and always \code{1} for Windows.}

\item{nbrCoresDiffMeth}{a positive \code{integer}, the number of cores
to use for parallel differential methylation calculations.Parameter
used for both sites and tiles analysis. The parameter
corresponds to the \code{num.cores} parameter in the package
\code{methylKit}.
Default: \code{1} and always \code{1} for Windows.}

\item{minReads}{a positive \code{integer} Bases and regions having lower
coverage than this count are discarded. The parameter
correspond to the \code{lo.count} parameter in the package \code{methylKit}.}

\item{minMethDiff}{a positive \code{double} betwwen [0,100], the absolute
value of methylation percentage change between cases and controls. The
parameter correspond to the \code{difference} parameter in
the methylKit package. Default: \code{10}.}

\item{qvalue}{a positive \code{double} betwwen [0,1], the cutoff
for qvalue of differential methylation statistic. Default: \code{0.01}.}

\item{maxPercReads}{a \code{double} between [0,100], the percentile of read
counts that is going to be used as upper cutoff. Bases ore regions
having higher
coverage than this percentile are discarded. Parameter used for both CpG
sites and tiles analysis. The parameter
correspond to the \code{hi.perc} parameter in the package \code{methylKit}.
Default: \code{99.9}.}

\item{destrand}{a \code{logical}, when \code{TRUE} will merge reads on both
strands of a CpG dinucleotide to provide better coverage. Only advised
when looking at CpG methylation. Parameter used for both CpG
sites and tiles analysis.
Default: \code{FALSE}.}

\item{minCovBasesForTiles}{a non-negative \code{integer}, the minimum
number of bases to be covered in a given tiling window. The parameter
corresponds to the \code{cov.bases} parameter in the package
\code{methylKit}.
Only used when \code{doingTiles} =
\code{TRUE}. Default: \code{0}.}

\item{tileSize}{a positive \code{integer}, the size of the tiling window.
The parameter corresponds to the \code{win.size} parameter in
the package \code{methylKit}. Only
used when \code{doingTiles} = \code{TRUE}. Default: \code{1000}.}

\item{stepSize}{a positive \code{integer}, the step size of tiling windows.
The parameter corresponds to the \code{stepSize} parameter in
the package \code{methylKit}. Only
used when \code{doingTiles} = \code{TRUE}. Default: \code{1000}.}

\item{vSeed}{a \code{integer}, a seed used when reproducible results are
needed. When a value inferior or equal to zero is given, a random integer
is used. Default: \code{-1}.}
}
\value{
@return a \code{list} containing the following elements:
\itemize{
\item \code{SITES} Only present when \code{type} = \code{"sites"} or
\code{"both"}, a \code{list} containing:
\itemize{
\item\code{i2} a \code{list} containing:
\itemize{
\item \code{HYPER} a \code{list} of \code{integer}, the number of conserved
hyper differentially methylated sites between two consecutive generations.
The first element represents the intersection of the first and second
generations; the second element, the intersection of the second and third
generations; etc..
\item \code{HYPO} a \code{list} of \code{integer}, the number of conserved
hypo differentially methylated sites between two consecutive generations.The
first element represents the intersection of the first and second
generations; the second element, the intersection of the second and third
generations; etc..
}
\item\code{iAll} a \code{list} containing:
\itemize{
\item \code{HYPER} a \code{list} of \code{integer}, the number of conserved
hyper differentially methylated sites between three or more consecutive
generations. The first element represents the intersection of the first
three generations; the second element, the intersection of the first fourth
generations; etc..The number of entries depends of the number
of generations.
\item \code{HYPO} a \code{list} of \code{integer}, the number of conserved
hypo differentially methylated sites between three or more consecutive
generations. The first element represents the intersection of the first
three generations; the second element, the intersection of the first fourth
generations; etc..The number of entries depends of the number of
generations.
}
}
\item \code{TILES} Only present when \code{type} = \code{"tiles"} or
\code{"both"}, a \code{list} containing:
\itemize{
\item\code{i2} a \code{list} containing:
\itemize{
\item \code{HYPER} a \code{list} of \code{integer}, the number of conserved
hyper differentially methylated positions between two consecutive
generations. The first element represents the intersection of the
first and second generations; the second element, the intersection of
the second and third generations; etc..
\item \code{HYPO} a \code{list} of \code{integer}, the number of conserved
hypo differentially methylated positions between two consecutive
generations.The first element represents the intersection of the first and
second generations; the second element, the intersection of the second
and third generations; etc..
}
\item\code{iAll} a \code{list} containing:
\itemize{
\item \code{HYPER} a \code{list} of \code{integer}, the number of conserved
hyper differentially methylated positions between three or more consecutive
generations. The first element represents the intersection of the first
three generations; the second element, the intersection of the first fourth
generations; etc..The number of entries depends of the number
of generations.
\item \code{HYPO} a \code{list} of \code{integer}, the number of conserved
hypo differentially methylated positions between three or more consecutive
generations. The first element represents the intersection of the first
three generations; the second element, the intersection of the first fourth
generations; etc..The number of entries depends of the number of
generations.
}
}
}
}
\description{
Run a differentially methylation analysis on each generation
present in a dataset. The number of conserved differentially
methylated elements (sites, tile or both) is them calculated. The
methylKit package is used to identify the differentially methylated
elements.
}
\examples{

## Path to a methylKit RDS file
methylFile <- dir(system.file("extdata", package = "methylInheritance"),
pattern = "methylObj_002.RDS", full.names = TRUE)

## Run a permutation analysis
\dontrun{runObservationUsingRDSFile(methylKitRDSFile = methylFile,
type = "sites", nbrPermutations = 0, minReads = 8, minMethDiff = 5,
vSeed = 2001)}

}
\author{
Astrid Deschenes, Pascal Belleau
}

