% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceMethods.R
\name{runPermutationUsingRDS}
\alias{runPermutationUsingRDS}
\title{Run all permutations on the specified multi-generational dataset in
RDS format.}
\usage{
runPermutationUsingRDS(methylKitRDSFile, type = c("both", "sites", "tiles"),
  outputDir = NULL, nbrPermutations = 1000, nbrCores = 1,
  nbrCoresDiffMeth = 1, minReads = 10, minMethDiff = 10, qvalue = 0.01,
  maxPercReads = 99.9, destrand = FALSE, minCovBasesForTiles = 0,
  tileSize = 1000, stepSize = 1000, vSeed = -1)
}
\arguments{
\item{methylKitRDSFile}{a string, the name of the file containing the
methylKit objet used for the permutation analysis.
TODO
of files with methylation information for
bases or regions in the genome. One \code{list} must contain all files
related to the same generation. So, if 3 generations are analyzed, a
\code{list} containing 3 \code{list} must be passed. At least 2 generations
must be present to do a permutation analysis.
The parameter
corresponds to the \code{location} parameter in the package \code{methylKit}.
A \code{list} of \code{vector} containing
\code{0} and \code{1}. The information indicating which files are
associated to controls (\code{0}) and which files are cases (\code{1}).
One \code{vector} must contain all informations
related to the same generation. So, if 3 generations are analyzed, a
\code{list} containing 3 \code{vector} must be passed. At least 2
generations
must be present to do a permutation analysis.
The parameter
corresponds to the \code{treatment} parameter in the package
\code{methylKit}.}

\item{type}{One of the "sites","tiles" or "both" strings. Specifies the type
of differentially methylated elements should be returned. For
retrieving differentially methylated bases type="sites"; for
differentially methylated regions type="tiles". Default: "both".}

\item{outputDir}{a string, the name of the directory that will contain
the results of the permutation or \code{NULL}. If the directory does not
exist, it will be created. When \code{NULL}, the results of the permutation
are not saved. Default: \code{NULL}.}

\item{nbrPermutations, }{a positive \code{integer}, the total number of
permutations that is going to be done. Default: \code{1000}.}

\item{nbrCores}{a positive \code{integer}, the number of cores to use when
processing the analysis. Default: \code{1} and always \code{1} for Windows.}

\item{nbrCoresDiffMeth}{a positive \code{integer}, the number of cores
to use for parallel differential methylation calculations.Parameter
used for both sites and tiles analysis. The parameter
corresponds to the \code{num.cores} parameter in the package
\code{methylKit}.
Default: \code{1} and always \code{1} for Windows.}

\item{minReads}{a positive \code{integer} Bases and regions having lower
coverage than this count are discarded. The parameter
correspond to the \code{lo.count} parameter in the package \code{methylKit}.}

\item{minMethDiff}{a positive \code{double} betwwen [0,100], the absolute
value of methylation percentage change between cases and controls. The
parameter correspond to the \code{difference} parameter in
the package \code{methylKit}. Default: \code{10}.}

\item{qvalue}{a positive \code{double} betwwen [0,1], the cutoff
for qvalue of differential methylation statistic. Default: \code{0.01}.}

\item{maxPercReads}{a \code{double} between [0,100], the percentile of read
counts that is going to be used as upper cutoff. Bases ore regions
having higher
coverage than this percentile are discarded. Parameter used for both CpG
sites and tiles analysis. The parameter
correspond to the \code{hi.perc} parameter in the package \code{methylKit}.
Default: \code{99.9}.}

\item{destrand}{a \code{logical}, when \code{TRUE} will merge reads on both
strands of a CpG dinucleotide to provide better coverage. Only advised
when looking at CpG methylation. Parameter used for both CpG
sites and tiles analysis.
Default: \code{FALSE}.}

\item{minCovBasesForTiles}{a non-negative \code{integer}, the minimum
number of bases to be covered in a given tiling window. The parameter
corresponds to the \code{cov.bases} parameter in the package
\code{methylKit}.
Only used when \code{doingTiles} =
\code{TRUE}. Default: \code{0}.}

\item{tileSize}{a positive \code{integer}, the size of the tiling window.
The parameter corresponds to the \code{win.size} parameter in
the package \code{methylKit}. Only
used when \code{doingTiles} = \code{TRUE}. Default: \code{1000}.}

\item{stepSize}{a positive \code{integer}, the step size of tiling windows.
The parameter corresponds to the \code{stepSize} parameter in
the package \code{methylKit}. Only
used when \code{doingTiles} = \code{TRUE}. Default: \code{1000}.}

\item{vSeed}{a \code{integer}, a seed used when reproducible results are
needed. When a value inferior or equal to zero is given, a random integer
is used. Default: \code{-1}.}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## Loading dataset
data(samplesForTransgenerationalAnalysis)

## Run a permutation analysis
\dontrun{runPermutationUsingRDS(methylKitRDSFile=samplesForTransgenerationalAnalysis,
type = "sites", nbrPermutations = 10, vSeed = 2001)}

}
\author{
Astrid Deschenes, Pascal Belleau
}

