% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceInternalMethods.R
\name{validateRunOnePermutation}
\alias{validateRunOnePermutation}
\title{Extract sample name from file name}
\usage{
validateRunOnePermutation(info, output_dir, genomeVersion, minReads, minCGs,
  tileSize, stepSize, minMethDiff, destrand, doingSites, doingTiles, debug)
}
\arguments{
\item{info}{TODO}

\item{output_dir}{TODO}

\item{genomeVersion}{TODO}

\item{minReads}{TODO}

\item{doingSites}{TODO}

\item{doingTiles}{TODO}

\item{debug}{TODO}
}
\value{
\code{0} indicating that all parameters validations have been
successful.
}
\description{
Extract sample name from file name
}
\examples{

##TODO

}
\author{
Astrid Deschenes, Pascal Belleau
}
\keyword{internal}

