% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceInternalMethods.R
\name{validateRunPermutation}
\alias{validateRunPermutation}
\title{Extract sample name from file name}
\usage{
validateRunPermutation(allFilesByGeneration, conditionsByGeneration, nbrCores,
  minReads, minCGs, tileSize, stepSize, minMethDiff, destrand, genomeVersion,
  nbrPermutations, output_dir, doingTiles, doingSites, vSeed)
}
\arguments{
\item{allFilesByGeneration}{TODO}

\item{conditionsByGeneration}{TODO}

\item{nbrCores}{a positive integer, the number of cores to use when
processing the analysis. Always \code{1} for Windows.}

\item{minCGs}{a positive integer, the minimum number of bases to
be covered in a given tiling window.}

\item{tileSize}{a positive integer, the size of the tiling window.}

\item{stepSize}{a positive integer, the step size of tiling windows.}

\item{minMethDiff}{a positive integer, the step size of tiling windows.}

\item{destrand}{a logical, when \code{TRUE} will merge reads on both
strands of a CpG dinucleotide to provide better coverage. Only advised
when looking at CpG methylation.}

\item{doingTiles}{a logical, when \code{TRUE} will do the analysis on the
tiles.}

\item{doingSites}{a logical, when \code{TRUE} will do the analysis on the
CpG dinucleotide sites.}

\item{vSeed}{a positive integer, either \code{-1} or a positive integer.}
}
\value{
\code{0} indicating that all parameters validations have been
successful.
}
\description{
Extract sample name from file name
}
\examples{

##TODO

}
\author{
Astrid Deschenes
}
\keyword{internal}

