% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceInternalMethods.R
\name{createDataStructure}
\alias{createDataStructure}
\title{TODO}
\usage{
createDataStructure(interGenerationResult)
}
\arguments{
\item{interGenerationResult}{TODO}

\item{type}{One of the "sites" or "tiles" strings. Specifies the type
of elements that should be returned. For
retrieving differentially methylated bases type="sites"; for
differentially methylated regions type="tiles". Default: "sites".}

\item{result}{TODO}
}
\value{
a \code{list} containing:
}
\description{
TODO
}
\examples{

##TODO

}
\author{
Astrid Deschenes, Pascal Belleau
}
\keyword{internal}

