% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceInternalMethods.R
\name{createOutputDir}
\alias{createOutputDir}
\title{Create directories that will contained the results of the
permutations in RDS format.}
\usage{
createOutputDir(output_dir, doingSites = TRUE, doingTiles = FALSE)
}
\arguments{
\item{output_dir}{a string of \code{character}, the name of the main
directory to be created.}

\item{doingSites}{a \code{logical}, TODO}

\item{doingTiles}{a \code{logical}, TODO}
}
\value{
\code{0} when all directories are created without problem.
}
\description{
Create directories that will contained the results of the
permutations in RDS format.
}
\examples{

## TODO

}
\author{
Astrid Deschenes
}
\keyword{internal}

