% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritance.R
\docType{data}
\name{demoForTransgenerationalAnalysis}
\alias{demoForTransgenerationalAnalysis}
\title{Methylation information from samples over three generations. Information
for each generation is stored in a
\code{methylRawList} format (for demo purpose).}
\format{A \code{list} containing three \code{methylRawList} objects. Each
\code{methylRawList} contains the information for one generation
(first entry = first generation, etc..). Each sample information is
stored in a \code{methylRaw} object. There is \code{methylRaw} objects
(6 controls and 6 cases) in each generation.}
\usage{
data(demoForTransgenerationalAnalysis)
}
\value{
A \code{list} containing three \code{methylRawList} objects. Each
\code{methylRawList} contains the information for one generation
(first entry = first generation, etc..). Each sample information is
stored in a \code{methylRaw} object. There is \code{methylRaw} objects
(6 controls and 6 cases) in each generation.
}
\description{
The object is a \code{list} with 3 entries. Each entry correspond to the
information for one generation (first entry = first generation, etc..)
stored in a \code{methylRawList} ojbect.
There is 12 samples (6 controls and 6 cases) for each generation. Each
sample information is stored in a \code{methylRaw} object.
}
\details{
This dataset can be
used to test \code{runPermutationUsingMethylKitInfo} and
{runObservationUsingMethylKitInfo} functions.
}
\examples{

## Loading dataset
data(demoForTransgenerationalAnalysis)

## Run a permutation analysis
\dontrun{(runObservationUsingMethylKitInfo(methylKitInfo =
demoForTransgenerationalAnalysis, type = "tiles", nbrPermutations = 3,
vSeed = 2001)}

}
\seealso{
\itemize{
    \item \code{\link{runPermutationUsingMethylKitInfo}} {for running a
    permutation analysis using methylKit info entry}
    \item \code{\link{runObservationUsingMethylKitInfo}} {for running a
    observation analysis using methylKit info entry}
}
}
\keyword{datasets}

