% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceMethods.R
\name{formatForGraph}
\alias{formatForGraph}
\title{TODO}
\usage{
formatForGraph(analysisandPermutationResults, type = c("sites", "tiles"),
  inter = c("i2", "iAll"), position)
}
\arguments{
\item{analysisandPermutationResults}{TODO}

\item{type}{TODO}

\item{inter}{TODO}

\item{position}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## TODO

}
\author{
Astrid Deschenes, Pascal Belleau
}

