% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceInternalMethods.R
\name{formatInputMethylData}
\alias{formatInputMethylData}
\title{Permute dataset}
\usage{
formatInputMethylData(methylKitData)
}
\arguments{
\item{methylKitData}{a \code{list} of \code{methylRawList} entries. Each
\code{methylRawList} entry must contain all the \code{methylRaw} entries
related to one generation (first entry = first generation, second
entry = second generation, etc..). The number of generations must
correspond to the number
of entries in the \code{methylKitData}. At least 2 generations
must be present to make a permutation analysis. More information can be
found in the methylKit package.}
}
\value{
a \code{list} of \code{methylRawList} entries.
}
\description{
Permute dataset and format it to be ready for an analysis
}
\examples{

## Load dataset
data("samplesForTransgenerationalAnalysis")

methylInheritance:::formatInputMethylData(samplesForTransgenerationalAnalysis)

}
\author{
Astrid Deschenes, Pascal Belleau
}
\keyword{internal}
