% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceInternalMethods.R
\name{getGRangesFromMethylDiff}
\alias{getGRangesFromMethylDiff}
\title{TODO}
\usage{
getGRangesFromMethylDiff(methDiff, pDiff, qCut, typeD = "all")
}
\arguments{
\item{methDiff, }{TODO}

\item{pDiff, }{TODO}

\item{qCut, }{TODO}

\item{typeD, }{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## TODO

}
\author{
Pascal Belleau
}
\keyword{internal}

