% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceInternalMethods.R
\name{getSampleNameFromFileName}
\alias{getSampleNameFromFileName}
\title{Extract sample name from file name}
\usage{
getSampleNameFromFileName(fileName)
}
\arguments{
\item{fileName}{a string, the file name used to extract the sample name.}
}
\value{
A sample name extracted from the specified file name.
}
\description{
Extract a sample name from the file name. The sample name
corresponds to  the file name but without path information.
}
\examples{

## Extract sample name from file name
methylInheritance:::getSampleNameFromFileName("./test/experiment001/file_J1.txt")

}
\author{
Astrid Deschenes
}
\keyword{internal}

