% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceInternalMethods.R
\name{interGeneration}
\alias{interGeneration}
\title{TODO}
\usage{
interGeneration(resultAllGenGR)
}
\arguments{
\item{resultAllGenGR}{\code{GRanges} from \code{getGRangesFromMethylDiff}, TODO}
}
\value{
\code{list} with 2 elements
        i2 list of intersection G1 and G2, G2 and G3, ...
        iAll list of intersection G1 and G2 and G3,
             G1 and G2 and G3 and G4, ...
        TODO
}
\description{
TODO
}
\examples{
## sum(width(res))
## TODO

}
\author{
Pascal Belleau
}
\keyword{internal}

