% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceMethods.R
\name{loadAllPermutationRDS}
\alias{loadAllPermutationRDS}
\title{Load all RDS files created by the permutation analysis}
\usage{
loadAllPermutationRDS(analysisResultsDIR, permutationResultsDIR,
  doingSites = TRUE, doingTiles = FALSE)
}
\arguments{
\item{analysisResultsDIR}{TODO}

\item{permutationResultsDIR}{TODO}

\item{doingSites}{TODO}

\item{doingTiles}{TODO}
}
\value{
an \code{list} of \code{class} "methylInheritanceAllResults"
containing the following elements:
\itemize{
    \item \code{OBSERVED}, a \code{list} that contains one or two
    \code{list} depending
    of the specified parameters. When \code{doingSites} = \code{TRUE}, a
    list of called "SITES" is present. When \code{doingTiles} = \code{TRUE},
    a list of called "TILES" is present. TODO
    \item \code{PERMUTATION}, a \code{list}  contains one or two \code{list}
    depending of the specified parameters. When
    \code{doingSites} = \code{TRUE}, a
    list of called "SITES" is present. When \code{doingTiles} = \code{TRUE},
    a list of called "TILES" is present. TODO
}
}
\description{
Load all RDS files created by the permutation analysis and
create an object of \code{class} "methylInheritanceAllResults" to hold all
the pertinent information.
}
\examples{

## TODO

}
\author{
Astrid Deschenes, Pascal Belleau
}

