% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceMethods.R
\name{loadAllRDSResults}
\alias{loadAllRDSResults}
\title{Load all RDS files created by the permutation  and observation
analysis}
\usage{
loadAllRDSResults(analysisResultsDir, permutationResultsDir,
  doingSites = TRUE, doingTiles = FALSE)
}
\arguments{
\item{analysisResultsDir}{a \code{character} string, the path to the
directory that contains the analysis results. The path can be the same that
for the \code{permutatioNResultsDir} parameter.}

\item{permutationResultsDir}{a \code{character} string, the path to the
directory that contains the permutation results. The path can be the same
that for the \code{analysisResultsDir} parameter.}

\item{doingSites}{a \code{logical}, the data related to differentially
methylated sites are loaded when
\code{doingSites} = \code{TRUE}. Default: \code{TRUE}.}

\item{doingTiles}{a \code{logical}, the data related to differentially
methylated tiles are loaded when
\code{doingTiles} = \code{TRUE}. Default: \code{TRUE}.}
}
\value{
an \code{list} of \code{class} "methylInheritanceAllResults"
containing the following elements:
\itemize{
    \item \code{OBSERVATION}, a \code{list} that contains one or two
    \code{list} depending
    of the specified parameters. When \code{doingSites} = \code{TRUE}, a
    list of called "SITES" is present. When \code{doingTiles} = \code{TRUE},
    a list of called "TILES" is present. TODO
    \item \code{PERMUTATION}, a \code{list}  contains one or two \code{list}
    depending of the specified parameters. When
    \code{doingSites} = \code{TRUE}, a
    list of called "SITES" is present. When \code{doingTiles} = \code{TRUE},
    a list of called "TILES" is present. TODO
}
}
\description{
Load all RDS files created by the permutation and
observation analysis. The function
returns an object of \code{class} "methylInheritanceAllResults" that holds
all the pertinent information.
}
\examples{

## Get the name of the directory where files are stored
filesDir <- dir(system.file("extdata", package = "methylInheritance"),
pattern = "TEST", full.names = TRUE)

## Load information from files
results <- loadAllRDSResults(analysisResultsDir = filesDir,
permutationResultsDir = filesDir, doingSites = TRUE, doingTiles = TRUE)

}
\author{
Astrid Deschenes, Pascal Belleau
}

