% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceMethods.R
\name{mergePermutationAndObservation}
\alias{mergePermutationAndObservation}
\title{Merge the permutation results with the observation results.}
\usage{
mergePermutationAndObservation(permutationResults, observationResults)
}
\arguments{
\item{permutationResults}{a \code{list} with 1 entry called
\code{PERMUTATION}m the entry is a \code{list} with a number of
entries corresponding
to the number of permutations that have been processed. Each entry contains
the result of one permutation.}

\item{observationResults}{a \code{list} with 1 entry called
\code{OBSERVATION}, the entry is a \code{list} of 1 entry containing
the results obtained
with the real dataset (not permutated).}
}
\value{
a \code{list} of class \code{} with 2 entries. The 2 entries are:
\itemize{
\item \code{PERMUTATION} \code{list} with a number of entries corresponding
to the number of permutations that have been processed. Each entry contains
the result of one permutation.
\item \code{OBSERVATION} a \code{list} with 1 entry, the results obtained
with the real dataset (not permutated).
}
}
\description{
Merge the permutation results with the observation results.
The merging is only needed when permutation and observation have been
processed separatly. The returned value is a
\code{methylInheritanceAllResults} object that can be used by
the \code{extractInfo} function.
}
\examples{

## TODO

}
\author{
Astrid Deschenes, Pascal Belleau
}

