% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritance.R
\docType{package}
\name{methylInheritance-package}
\alias{methylInheritance}
\alias{methylInheritance-package}
\title{methylInheritance: Permutation}
\description{
This package does a permutation analysis, based on Monte Carlo sampling,
for testing the hypothesis that the number of conserved differentially
methylated elements (sites, tiles or regions), between
several generations, is associated to an effect inherited from a treatment
and that stochastic effect can be dismissed.
}
\author{
Astrid Deschênes,
Pascal Belleau and
Arnaud Droit

Maintainer:
Astrid Deschenes <astrid-louise.deschenes@crchudequebec.ulaval.ca>
}
\seealso{
\itemize{
    \item \code{\link{runPermutationUsingRDSFile}} { for running a
    permutation analysis on the specified multi-generational dataset in
    RDS format}
}
}
\keyword{package}

