% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceMethods.R
\name{plotGraph}
\alias{plotGraph}
\title{TODO}
\usage{
plotGraph(formatForGraphDataFrame)
}
\arguments{
\item{formatForGraphDataFrame}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\author{
Astrid Deschenes, Pascal Belleau
}

