% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceInternalMethods.R
\name{printDiffMethylFile}
\alias{printDiffMethylFile}
\title{TODO}
\usage{
printDiffMethylFile(data, output_dir, count, designName, isHyper, isSites)
}
\arguments{
\item{data}{TODO}

\item{output_dir}{TODO}

\item{count}{TODO}

\item{designName}{TODO}

\item{isHyper}{TODO}

\item{isSites}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## TODO


}
\author{
Astrid Deschenes
}
\keyword{internal}

