% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceInternalMethods.R
\name{runOnePermutation}
\alias{runOnePermutation}
\title{TODO}
\usage{
runOnePermutation(info, output_dir, doingSites = TRUE, doingTiles = FALSE,
  debug = FALSE)
}
\arguments{
\item{info}{TODO}

\item{doingSites}{Default: \code{TRUE}.}

\item{doingTiles}{Default: \code{FALSE}.}

\item{debug}{Default: \code{FALSE}.}

\item{output_di}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

##TODO

}
\author{
Astrid Deschenes, Pascal Belleau
}
\keyword{internal}

