% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceMethods.R
\name{runPermutation}
\alias{runPermutation}
\title{TODO}
\usage{
runPermutation(allFilesByGeneration, conditionsByGeneration, nbrCores = 1,
  minReads = 10, minCGs, tileSize, stepSize, minMethDiff, destrand,
  genomeVersion, nbrPermutations = 1000, output_dir, doingTiles, doingSites,
  vSeed = -1)
}
\arguments{
\item{allFilesByGeneration}{TODO}

\item{conditionsByGeneration}{TODO}

\item{nbrCores}{a positive integer, the number of cores to use when
processing the analysis. Default: \code{1} and always \code{1} for Windows.}

\item{minReads}{TODO}

\item{minCGs}{TODO}

\item{tileSize}{TODO}

\item{stepSize}{TODO}

\item{minMethDiff}{TODO}

\item{destrand}{TODO}

\item{genomeVersion}{TODO}

\item{nbrPermutations}{TODO}

\item{output_dir}{TODO

 @param doingSites a logical, when \code{TRUE} will do the analysis on the
CpG dinucleotide sites.}

\item{doingTiles}{a logical, when \code{TRUE} will do the analysis on the
tiles.}

\item{vSeed}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

##TODO

}
\author{
Astrid Deschenes, Pascal Belleau
}

