% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceMethods.R
\name{runPermutation}
\alias{runPermutation}
\title{TODO}
\usage{
runPermutation(allDataByGeneration, infoByGeneration, nbCores = 1,
  nbrPermutations = 1000, output_dir, vSeed = -1)
}
\arguments{
\item{allDataByGeneration}{TODO}

\item{infoByGeneration}{TODO}

\item{nbCores}{Default = 1.}

\item{nbrPermutations}{TODO}

\item{output_dir}{TODO}

\item{vSeed}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

##TODO

}
\author{
Astrid Deschenes, Pascal Belleau
}

