% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceMethods.R
\name{runPermutation}
\alias{runPermutation}
\title{TODO}
\usage{
runPermutation(allFilesByGeneration, conditionsByGeneration, nbCores = 1,
  minReads, minCGs, tileSize, stepSize, minMethDiff, mergeStrand, genomeVersion,
  nbrPermutations = 1000, output_dir, doingTiles, doingSites, vSeed = -1)
}
\arguments{
\item{allFilesByGeneration}{TODO}

\item{conditionsByGeneration}{TODO}

\item{nbCores}{Default = 1.}

\item{minReads}{TODO}

\item{minCGs}{TODO}

\item{tileSize}{TODO}

\item{stepSize}{TODO}

\item{minMethDiff}{TODO}

\item{mergeStrand}{TODO}

\item{genomeVersion}{TODO}

\item{nbrPermutations}{TODO}

\item{output_dir}{TODO}

\item{doingTiles}{TODO}

\item{doingSites}{TODO}

\item{vSeed}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

##TODO

}
\author{
Astrid Deschenes, Pascal Belleau
}

