% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceInternalMethods.R
\name{saveInterGenerationResults}
\alias{saveInterGenerationResults}
\title{Run one permutation using \code{methylKit} package. One permutation
includes analysis for all generations associated to the same permutation.}
\usage{
saveInterGenerationResults(outputDir, permutationID, type = c("sites",
  "tiles"), result)
}
\arguments{
\item{outputDir}{a string, the name of the directory that will contain
the results of the permutation. The name should end with a slash.}

\item{permutationID}{an integer, the identifiant of the permutation.}

\item{type}{One of the "sites","tiles" or "both" strings. Specifies the type
of differentially methylated elements should be returned. For
retrieving differentially methylated bases type="sites"; for
differentially methylated regions type="tiles". Default: "both".}

\item{result}{TODO}

\item{type}{One of the "sites" or "tiles" strings. Specifies the type
of differentially methylated elements should be saved. Default: "sites".}
}
\value{
\code{0} indicating that all parameters validations have been
successful.
}
\description{
Run one CpG site or region analysis using the \code{methylKit}
package. The output of the analysis is saved in a file in the specified
directory.
}
\examples{

##TODO

}
\author{
Astrid Deschenes, Pascal Belleau
}
\keyword{internal}

